+++ test-index.js
@@ -1,0 +1,1 @@
+document.getElementById("root").innerHTML += "hi";
